Add config for VSCode integration
This codebase reflects my own workflow and tooling choices. In order for others to find it easier to contribute a list of recommended plugins and settings has been added.

The suggested plugins should make it easier to read and write code without false-positive syntax violations being reported by the.

Config for eslint and stylelint is found under `./tools`: the `.vscode/settings.json` tells VSCode where to find them so that developers get instant feedback rather than only when the linters are run (e.g. on commit)